+++ src/hopkins_statistic/_inference.py
@@ -68,6 +68,7 @@ def hopkins_test(
         frame: Area sampling frame. Must be one of:
 
             - Literal `'bbox'` to use the axis-aligned bounding box of `X`, or
+            - Literal `'hull'` to use the convex hull of `X`, or
             - Pair `(lower, upper)` defining the bounds of a rectangular
               sampling frame. Both must be broadcastable to shape `(d,)`.
               While data points outside a given frame are ignored during